The monolithic   utils.kt has been split into thematic files (  DateUtils.kt,   SimUtils.kt, etc.), making helper functions easier to find and maintain. 2. Notification Management: All CallStyle and notification channel logic is now encapsulated in the new   CallNotificationManager, reducing the "God Service" complexity of CallService. 3. Architectural Decoupling: I introduced   ICallRepository to abstract call states. CallService no longer relies on static companion object flows, significantly improving testability. 4. Activity Simplification: MainActivity has been decluttered by moving intent processing and biometric lock logic to a dedicated   MainViewModel. 5. UI Performance: The monolithic CallScreen has been decomposed into smaller, reusable components in   CallScreenComponents.kt, which optimizes Jetpack Compose recomposition scopes.
