+++ app/src/main/java/dev/goodwy/rphone/controller/util/PermissionUtils.kt
@@ -1,0 +1,41 @@
+package dev.goodwy.rphone.controller.util
+
+import android.app.role.RoleManager
+import android.content.Context
+import android.content.Intent
+import android.content.pm.PackageManager
+import android.os.Build
+import android.telecom.TelecomManager
+
+fun isAlreadyDefaultDialer(context: Context): Boolean {
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+        val roleManager = context.getSystemService(Context.ROLE_SERVICE) as RoleManager
+        roleManager.isRoleHeld(RoleManager.ROLE_DIALER)
+    } else {
+        val telecomManager = context.getSystemService(Context.TELECOM_SERVICE) as TelecomManager
+        val defaultDialerPackage = telecomManager.defaultDialerPackage
+        defaultDialerPackage == context.packageName
+    }
+}
+
+fun getDefaultDialerIntent(context: Context): Intent {
+    return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+        val roleManager = context.getSystemService(Context.ROLE_SERVICE) as RoleManager
+        roleManager.createRequestRoleIntent(RoleManager.ROLE_DIALER)
+    } else {
+        Intent(TelecomManager.ACTION_CHANGE_DEFAULT_DIALER).apply {
+            putExtra(TelecomManager.EXTRA_CHANGE_DEFAULT_DIALER_PACKAGE_NAME, context.packageName)
+        }
+    }
+}
+
+fun Context.isPackageInstalled(packageName: String?): Boolean {
+    if (packageName == null) return false
+    return try {
+        val packageManager = packageManager
+        val intent = packageManager.getLaunchIntentForPackage(packageName)
+        intent != null && packageManager.queryIntentActivities(intent, PackageManager.MATCH_DEFAULT_ONLY).isNotEmpty()
+    } catch (e: Exception) {
+        false
+    }
+}
